+++ spotbugs/src/test/java/edu/umd/cs/findbugs/util/MutableClassesTest.java
@@ -100,4 +100,31 @@ public class MutableClassesTest {
     public void TestImmutable() {
         Assert.assertFalse(MutableClasses.mutableSignature("Ledu/umd/cs/findbugs/util/MutableClassesTest$Immutable;"));
     }
+
+    @Test
+    public void TestImmutableValuedBased() {
+        // Annotated with @jdk.internal.ValueBased and has "setValue", which should normally trip detection
+        Assert.assertFalse(MutableClasses.mutableSignature("Ljava/util/KeyValueHolder;"));
+    }
+
+    @com.google.errorprone.annotations.Immutable
+    public static class ErrorProneImmutable {
+        public void write() {
+            // Does not matter
+        }
+    }
+
+    public static class ErrorProneImmutableSubclass extends ErrorProneImmutable {
+        public void writeOther() {
+            // Does not matter
+        }
+    }
+
+    @Test
+    public void TestErrorProneImmutable() {
+        Assert.assertFalse(MutableClasses.mutableSignature(
+                "Ledu/umd/cs/findbugs/util/MutableClassesTest$ErrorProneImmutable;"));
+        Assert.assertFalse(MutableClasses.mutableSignature(
+                "Ledu/umd/cs/findbugs/util/MutableClassesTest$ErrorProneImmutableSubclass;"));
+    }
 }